requirements.txt
# Core dependencies
asyncpg>=0.29.0
python-dateutil>=2.8.0

# Meteorological data processing
eccodes>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Minimal geospatial (only what's actually needed)
shapely>=2.0.0

# HTTP client
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0